Remove user_identifiers table from being synced
Because user_identifiers draws their data from a foreign key
relationship, and it's unclear if the syncing service can traverse that,
rather don't sync that table (for now). It will have implications for
the school admin of NA but this may not be an issue and may really just
prompt a re-factoring to remove this table for good.